// src/interfaces/appInterfaces.tsx
// Generated by https://quicktype.io

export interface LoginData {
    correo: string;
    password: string;
}

export interface RegisterData {
    nombre: string;
    correo: string;
    password: string;
}


export interface LoginResponse {
    usuario: Usuario;
    token:   string;
}

export interface Usuario {
    rol:    string;
    estado: boolean;
    google: boolean;
    nombre: string;
    correo: string;
    uid:    string;
    img?: string;
}

// Generated by https://quicktype.io

export interface ResponseError {
    errors: Error[];
}

export interface Error {
    value:    string;
    msg:      string;
    param:    string;
    location: string;
}